Introducing an important difference between jmsConfig parameters and messageProperties: Where as JMS allows to send a (subset) of java primitive types, for the configuration of the JMS connection we need strings only (for the time being).

+++ java/src/org/exist/messaging/configuration/MessagingConfiguration.java
@@ -78,7 +78,7 @@ public class MessagingConfiguration extends Properties {
 
             } else if (values instanceof IntegerValue) {
                 IntegerValue singleValue = (IntegerValue) values;
-                put(keyValue, singleValue.toJavaObject(BigInteger.class));
+                put(keyValue, singleValue.toJavaObject(Integer.class));
 
             } else if (values instanceof DoubleValue) {
                 DoubleValue singleValue = (DoubleValue) values;
